# api/openapi.yaml
openapi: 3.0.0
info:
  title: OASMock control HTTP API
  description: HTTP API for the OpenAPI-based mock tool
  version: 0.1.0
servers:
  - url: http://localhost:19191/_mock
    description: Default mock server
paths:
...
  /ws/push:
    post:
      operationId: pushToChannel
      summary: Push a message to channel consumers
      description: |
        Pushes a message to the connected consumers of an AsyncAPI channel.
        Supports immediate or delayed delivery, targeted (by connectionId) or
        broadcast push, and runtime-expression templating of the payload.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushRequest'
      responses:
        '200':
          description: Push accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncActionResponse'
        '400':
          description: Invalid request
        '404':
          description: Unknown connectionId
...
components:
  schemas:
...
    PushRequest:
      type: object
      required:
        - channel
      properties:
        channel:
          type: string
          description: AsyncAPI channel address to push to
        connectionId:
          type: string
          description: Target a specific consumer; omitted broadcasts to all
        payload:
          type: object
          description: Message payload, templated with {$state.*}/{$env.*}
        delay:
          type: integer
          minimum: 0
          default: 0
          description: Delivery delay in milliseconds
...
    AsyncActionResponse:
      type: object
      properties:
        success:
          type: boolean
        event:
          type: string
          description: Fired event name (fire-event endpoint only)
        pushId:
          type: string
          description: Scheduled push id (schedule endpoint only)
...